Make some initial attempt at `get_datetimes_from_filters`
And made EventStream build the URL instead of the backend. This makes it
easier in the future to be interactive with the event stream, and
additionally, since we know if it's a stream or not, we can call
`.close()` manually, which for some reason seem neccessary (if this is
not done, it will do the same request in an interval instead..) There's
probably better ways to solve that, though. (Like sending a quit-event
or similar.)